Use maximum available native integer type

+++ main.cpp
@@ -1,7 +1,8 @@
 #include <iostream>
+#include <cstdint>
 
 int main() {
-    int a, b;
+    intmax_t a, b;
     std::cin >> a >> b;
     std::cout << a + b << std::endl;
     return 0;
